~ fixed PyList array-like accessor not allowing sets ~ changed PyDictionary to allow PyDataType keys instead of only strings ~ added the same < and > operatos to PyDecial as we did with PyInteger + added support for Indexed Rowsets + Add PyItemInfo and PyWarInfo objects to ease the usage of the correspondent Indexed Rowsets ~ temporary fix for PyPackedRow not being properly decoded (not that important, here for completeness sake, mostly) + added support for BoundServices (YAY!) + added BoundServiceManager + Add station service and station::GetStationItemBits call + Added warRegistry Bound service and warRegistry::GetWars + Added invbroker service and invbroker::GetInventoryFromId + Added dogmaIM service and dogmaIM::ShipGetInfo + Added config::GetMultiGraphicsEx + Added character::SelectCharacterID with the proper session change notification + Added support for BoundService calls on the Cluster Controller + Added StationManager class + Added Station class + Added Ship class ~ Items are back to using IDs for location and owner, prevents a recursive call to LoadItem on the ItemManager, these should be loaded on-demand + First steps into Inventory support + Important changes to the database structure, there's still a lot of changes pending, but this should suffice for now, please re-build your whole database from scratch + Added extra session data to the Client class so It can be accessed at any time ~ updated some of the queries to the database for generating the cache to take advantage of the new database changes + added some dataID fields missing in various caches, still more to go ~ normal service manager will only call public and instance methods + added a missing modification on PingReq

+++ Server/PythonTypes/Types/Database/Utils.cs
@@ -31,15 +31,15 @@ namespace PythonTypes.Types.Database
             if (type == typeof(string))
                 data = new PyString(reader.GetString(index), true);
             else if (type == typeof(ulong))
-                data = reader.GetUInt64(index);
+                data = (long) reader.GetUInt64(index);
             else if (type == typeof(long))
                 data = reader.GetInt64(index);
             else if (type == typeof(uint))
-                data = reader.GetUInt32(index);
+                data = (int) reader.GetUInt32(index);
             else if (type == typeof(int))
                 data = reader.GetInt32(index);
             else if (type == typeof(ushort))
-                data = reader.GetUInt16(index);
+                data = (short) reader.GetUInt16(index);
             else if (type == typeof(short))
                 data = reader.GetInt16(index);
             else if (type == typeof(byte))
